+++ src/CommonRuntime/StructuralInference.fs
@@ -1,4 +1,4 @@
-﻿/// Implements type inference for unstructured documents like XML or JSON
+/// Implements type inference for unstructured documents like XML or JSON
 module FSharp.Data.Runtime.StructuralInference
 
 open System
@@ -31,7 +31,7 @@ module List =
 let private numericTypes = [ typeof<Bit0>; typeof<Bit1>; typeof<int>; typeof<int64>; typeof<decimal>; typeof<float>]
 
 /// List of primitive types that can be returned as a result of the inference
-let private primitiveTypes = [typeof<string>; typeof<DateTime>; typeof<DateTimeOffset>; typeof<Guid>; typeof<bool>; typeof<Bit>] @ numericTypes
+let private primitiveTypes = [typeof<string>; typeof<DateTime>; typeof<DateTimeOffset>; typeof<TimeSpan>; typeof<Guid>; typeof<bool>; typeof<Bit>] @ numericTypes
 
 /// Checks whether a type supports unit of measure
 let supportsUnitsOfMeasure typ =    
@@ -49,6 +49,7 @@ let typeTag = function
       elif typ = typeof<bool> then InferedTypeTag.Boolean
       elif typ = typeof<string> then InferedTypeTag.String
       elif typ = typeof<DateTime> || typ = typeof<DateTimeOffset> then InferedTypeTag.DateTime
+      elif typ = typeof<TimeSpan> then InferedTypeTag.TimeSpan
       elif typ = typeof<Guid> then InferedTypeTag.Guid
       else failwith "typeTag: Unknown primitive type"
   | InferedType.Json _ -> InferedTypeTag.Json
@@ -233,7 +234,7 @@ module private Helpers =
     |> Seq.length
 
 /// Infers the type of a simple string value
-/// Returns one of null|typeof<Bit0>|typeof<Bit1>|typeof<bool>|typeof<int>|typeof<int64>|typeof<decimal>|typeof<float>|typeof<Guid>|typeof<DateTime>|typeof<string>
+/// Returns one of null|typeof<Bit0>|typeof<Bit1>|typeof<bool>|typeof<int>|typeof<int64>|typeof<decimal>|typeof<float>|typeof<Guid>|typeof<DateTime>|typeof<TimeSpan>|typeof<string>
 let inferPrimitiveType (cultureInfo:CultureInfo) (value : string) =
 
   // Helper for calling TextConversions.AsXyz functions
@@ -264,6 +265,7 @@ let inferPrimitiveType (cultureInfo:CultureInfo) (value : string) =
   | ParseNoCulture TextConversions.AsBoolean _ -> typeof<bool>
   | Parse TextConversions.AsInteger _ -> typeof<int>
   | Parse TextConversions.AsInteger64 _ -> typeof<int64>
+  | Parse TextConversions.AsTimeSpan _ -> typeof<TimeSpan>
   | Parse TextConversions.AsDateTimeOffset dateTimeOffset when not (isFakeDate dateTimeOffset.UtcDateTime value) -> typeof<DateTimeOffset>
   | Parse TextConversions.AsDateTime date when not (isFakeDate date value) -> typeof<DateTime>
   | Parse TextConversions.AsDecimal _ -> typeof<decimal>