+++ src/Html/HtmlParser.fs
@@ -1,4 +1,4 @@
-﻿#nowarn "10001"
+#nowarn "10001"
 namespace FSharp.Data
 
 open System
@@ -146,7 +146,10 @@ type HtmlNode =
     /// [omit]
     [<EditorBrowsableAttribute(EditorBrowsableState.Never)>]
     [<CompilerMessageAttribute("This method is intended for use in generated code only.", 10001, IsHidden=true, IsError=false)>]
-    member x._Print = x.ToString()
+    member x._Print =
+        let str = x.ToString()
+        if str.Length > 512 then str.Substring(0, 509) + "..."
+        else str
 
 [<StructuredFormatDisplay("{_Print}")>]
 /// Represents an HTML document
@@ -178,7 +181,11 @@ type HtmlDocument =
     /// [omit]
     [<EditorBrowsableAttribute(EditorBrowsableState.Never)>]
     [<CompilerMessageAttribute("This method is intended for use in generated code only.", 10001, IsHidden=true, IsError=false)>]
-    member x._Print = x.ToString()
+    member x._Print =
+        let str = x.ToString()
+        if str.Length > 512 then str.Substring(0, 509) + "..."
+        else str
+
 
 // --------------------------------------------------------------------------------------
 
@@ -329,7 +336,7 @@ module internal HtmlParser =
 
         member x.IsScriptTag 
             with get() = 
-               match x.CurrentTagName() with
+               match x.CurrentTagName().Trim().ToLower() with
                | "script" | "style" -> true
                | _ -> false
 
@@ -445,7 +452,7 @@ module internal HtmlParser =
             match state.Peek() with
             | '/' -> state.Cons(); scriptSingleLineComment state
             | '*' -> state.Cons(); scriptMultiLineComment state
-            | _ -> scriptRegex state
+            | _ -> script state
         and scriptMultiLineComment state =
             match state.Peek() with
             | TextParser.EndOfFile _ -> data state
@@ -461,15 +468,6 @@ module internal HtmlParser =
             | TextParser.EndOfFile _ -> data state
             | '\n' -> state.Cons(); script state
             | _ -> state.Cons(); scriptSingleLineComment state
-        and scriptRegex state =
-            match state.Peek() with
-            | TextParser.EndOfFile _ -> data state
-            | '/' -> state.Cons(); script state
-            | '\\' -> state.Cons(); scriptRegexBackslash state
-            | _ -> state.Cons(); scriptRegex state
-        and scriptRegexBackslash state =
-            match state.Peek() with
-            | _ -> state.Cons(); scriptRegex state
         and scriptLessThanSign state =
             match state.Peek() with
             | '/' -> state.Pop(); scriptEndTagOpen state
@@ -565,13 +563,12 @@ module internal HtmlParser =
             | TextParser.Whitespace _ -> state.Pop(); beforeAttributeName state
             | '/' when state.IsScriptTag -> state.Pop(); selfClosingStartTag state
             | '>' when state.IsScriptTag -> state.Pop(); state.EmitTag(true);
-            | '>' -> 
+            | TextParser.Letter _ -> state.ConsTag(); scriptEndTagName state
+            | _ ->
                 state.Cons([|'<'; '/'|]); 
                 state.Cons(state.CurrentTagName()); 
                 (!state.CurrentTag).Clear()
                 script state
-            | TextParser.Letter _ -> state.ConsTag(); scriptEndTagName state
-            | _ -> state.Cons('<'); state.Cons('/'); script state
         and charRef state = 
             match state.Peek() with
             | ';' -> state.Cons(); state.Emit()