+++ RELEASE_NOTES.md
@@ -1,8 +1,10 @@
-#### 3.0.0-beta5 - Unreleased
+#### 3.0.0-rc - October 7 2018
 * (Breaking Change) Added support for DateTimeOffset to JSON, XML, CSV, and HTML providers.
-* Map XSD date to System.DateTime and XSD dateTime to System.DateTimeOffset
+* (Breaking Change) Added support for TimeSpan to JSON, XML, CSV, and HTML providers.
+* Map XSD date to System.DateTime and XSD dateTime to System.DateTimeOffset.
 * Fixed large float values being silently converted to int.MinValue when parsing JSON.
 * Improved handling of invalid cookies.
+* Fixes for #1091 - tidying up regex handling in script tags.
 
 #### 3.0.0-beta4 - July 8 2018
 * (Breaking Change) Ignore culture when parsing JSON to better match the JSON spec.